backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic-settings==2.1.0
email-validator==2.1.1
